修改UserController位置
